Refactored course overview to make it more harmonious with other codes
