--- src/test/java/org/amadeus/charon/data/CourseOVPageTest.java
@@ -1,15 +1,0 @@
-package org.amadeus.charon.data;
-
-import junit.framework.TestCase;
-
-public class CourseOVPageTest extends TestCase {
-	
-	
-	public void isCourse(Course check){
-		assertTrue(check.getCourseCode() 
-		!= null && check.getCourseName() != null && check.getCourseDesc() != null);	
-	}
-	
-		
-	
-}
